// deploy_temp/src/types/dashboard.ts
export interface EmpresaCliente {
    id: number;
    rut: string;
    nombre: string;
    contacto: string;
    telefono: string;
    email: string;
    activo: boolean;
}